Shift UniswapV2 & UniswapV3 into new modules
