+++ src/OHMLUSDBondV1.ts
@@ -1,13 +1,13 @@
 import { DepositCall } from "../generated/OHMDAIBondV4/OHMDAIBondV4";
-import { toDecimal } from "./utils/Decimals";
 import { OHMLUSDLPBOND_TOKEN, PAIR_UNISWAP_V2_OHM_LUSD } from "./utils/Constants";
-import { loadOrCreateToken } from "./utils/Tokens";
 import { createDailyBondRecord } from "./utils/DailyBond";
-import { getOhmUSDPairValue } from "./utils/Price";
+import { toDecimal } from "./utils/Decimals";
+import { getOhmUSDPairValue } from "./utils/LiquidityUniswapV2";
+import { loadOrCreateToken } from "./utils/Tokens";
 
 export function handleDeposit(call: DepositCall): void {
-  let token = loadOrCreateToken(OHMLUSDLPBOND_TOKEN);
-  let amount = toDecimal(call.inputs._amount, 18);
+  const token = loadOrCreateToken(OHMLUSDLPBOND_TOKEN);
+  const amount = toDecimal(call.inputs._amount, 18);
 
   createDailyBondRecord(
     call.block.timestamp,
