+++ src/utils/Price.ts
@@ -15,7 +15,7 @@ import {
   PAIR_UNISWAP_V2_OHM_DAI_V2_BLOCK,
   PAIR_UNISWAP_V2_USDC_ETH,
 } from "./Constants";
-import { getERC20, getUniswapV2Pair, getUniswapV3Pair } from "./ContractHelper";
+import { getERC20 } from "./ContractHelper";
 import { toDecimal } from "./Decimals";
 import { PairHandlerTypes } from "./PairHandler";
 
@@ -246,7 +246,7 @@ export function getBaseTokenUSDRate(
  * @param pairAddress
  * @returns
  */
-function getUSDRateUniswapV2(
+export function getUSDRateUniswapV2(
   contractAddress: string,
   pairAddress: string,
   blockNumber: BigInt,
@@ -321,7 +321,7 @@ function getUSDRateUniswapV2(
   return baseTokenNumerator.times(baseTokenUsdRate);
 }
 
-function getUSDRateUniswapV3(
+export function getUSDRateUniswapV3(
   contractAddress: string,
   pairAddress: string,
   blockNumber: BigInt,
@@ -437,232 +437,4 @@ export function getUSDRate(contractAddress: string, blockNumber: BigInt): BigDec
       ") for contract: " +
       contractAddress,
   );
-}
-
-/**
- * To calculate the risk-free value of an OHM-DAI LP, we assume
- * that DAI = $1 and OHM = $1.
- *
- * The multiple of the quantity of tokens on both sides of the LP
- * remains constant in a Uniswap V2 pool: x * y = k
- *
- * Given this: x1 * y1 = x2 * y2
- *
- * However, if x2 = y2, then: x1 * y1 = x2^2
- *
- * x2 = sqrt(x1 * y1)
- *
- * This tells us the number of DAI (or OHM) tokens required at the
- * position on the constant product curve.
- *
- * If we assume that 1 DAI = 1 OHM, then the value of the entire
- * liquidity pool at RFC is: (1 + 1) * sqrt(# DAI * # OHM)
- *
- * The total value, given the balance, is therefore:
- *
- * (# LP tokens / LP total supply) * (2) * sqrt(# DAI * # OHM)
- *
- * This blog also helps illustrate it: https://olympusdao.medium.com/a-primer-on-oly-bonds-9763f125c124
- */
-export function getOhmUSDPairRiskFreeValue(
-  lpBalance: BigInt,
-  pairAddress: string,
-  blockNumber: BigInt,
-): BigDecimal {
-  // TODO assumes that the pair is Uniswap V2. What about V3 or balancer?
-  // TODO assumes part of the pair is a stablecoin
-  // TODO abstract out for ANY pair
-  const pair = getUniswapV2Pair(pairAddress, blockNumber);
-  if (!pair) {
-    throw new Error(
-      "Cannot determine discounted value as the contract " + pairAddress + " does not exist yet.",
-    );
-  }
-
-  const total_lp = pair.totalSupply();
-  const lp_token_1 = toDecimal(pair.getReserves().value0, 9);
-  const lp_token_2 = toDecimal(pair.getReserves().value1, 18);
-  const kLast = lp_token_1.times(lp_token_2).truncate(0).digits;
-
-  const part1 = toDecimal(lpBalance, 18).div(toDecimal(total_lp, 18));
-  const two = BigInt.fromI32(2);
-
-  const sqrt = kLast.sqrt();
-  const part2 = toDecimal(two.times(sqrt), 0);
-  const result = part1.times(part2);
-  log.debug("OHM-DAI risk-free value is {}", [result.toString()]);
-  return result;
-}
-
-export function getUniswapV3PairTotalValue(pairAddress: string, blockNumber: BigInt): BigDecimal {
-  const pair = getUniswapV3Pair(pairAddress, blockNumber);
-  if (!pair) {
-    throw new Error(
-      "Cannot determine discounted value as the contract " + pairAddress + " does not exist yet.",
-    );
-  }
-
-  // Determine token0 value
-  const token0 = pair.token0().toHexString();
-  log.debug("token0: {}", [token0]);
-  const token0Contract = getERC20(getContractName(token0), token0, blockNumber);
-  if (!token0Contract) {
-    throw new Error("Unable to find ERC20 contract for " + token0);
-  }
-
-  const token0Reserves = toDecimal(
-    token0Contract.balanceOf(Address.fromString(pairAddress)),
-    token0Contract.decimals(),
-  );
-  const token0Rate = getUSDRateUniswapV3(token0, pairAddress, blockNumber);
-  const token0Value = token0Reserves.times(token0Rate);
-  log.debug("token0: reserves = {}, rate = {}, value: {}", [
-    token0Reserves.toString(),
-    token0Rate.toString(),
-    token0Value.toString(),
-  ]);
-
-  // Determine token1 value
-  const token1 = pair.token1().toHexString();
-  log.debug("token1: {}", [token1]);
-  const token1Contract = getERC20(getContractName(token1), token1, blockNumber);
-  if (!token1Contract) {
-    throw new Error("Unable to find ERC20 contract for " + token1);
-  }
-
-  const token1Reserves = toDecimal(
-    token1Contract.balanceOf(Address.fromString(pairAddress)),
-    token1Contract.decimals(),
-  );
-  // Cheating, a little bit
-  const token1Rate = Address.fromString(token1).equals(Address.fromString(ERC20_WETH))
-    ? getBaseEthUsdRate()
-    : getUSDRateUniswapV3(token1, pairAddress, blockNumber);
-  const token1Value = token1Reserves.times(token1Rate);
-  log.debug("token1: reserves = {}, rate = {}, value: {}", [
-    token1Reserves.toString(),
-    token1Rate.toString(),
-    token1Value.toString(),
-  ]);
-
-  const pairValue = token0Value.plus(token1Value);
-  log.debug("UniswapV3 pair value for contract {} is: {}", [pairAddress, pairValue.toString()]);
-  return pairValue;
-}
-
-export function getUniswapV2PairTotalValue(pairAddress: string, blockNumber: BigInt): BigDecimal {
-  log.info("Calculating total value of pair {}", [pairAddress]);
-  const pair = getUniswapV2Pair(pairAddress, blockNumber);
-  if (!pair) {
-    throw new Error(
-      "Cannot determine discounted value as the contract " + pairAddress + " does not exist yet.",
-    );
-  }
-
-  // Determine token0 value
-  const token0 = pair.token0().toHexString();
-  log.debug("token0: {}", [token0]);
-  const token0Contract = getERC20(getContractName(token0), token0, blockNumber);
-  if (!token0Contract) {
-    throw new Error("Unable to find ERC20 contract for " + token0);
-  }
-
-  const token0Reserves = toDecimal(pair.getReserves().value0, token0Contract.decimals());
-  const token0Rate = getUSDRateUniswapV2(token0, pairAddress, blockNumber);
-  const token0Value = token0Reserves.times(token0Rate);
-  log.debug("token0: reserves = {}, rate = {}, value: {}", [
-    token0Reserves.toString(),
-    token0Rate.toString(),
-    token0Value.toString(),
-  ]);
-
-  // Determine token1 value
-  const token1 = pair.token1().toHexString();
-  log.debug("token1: {}", [token1]);
-  const token1Contract = getERC20(getContractName(token1), token1, blockNumber);
-  if (!token1Contract) {
-    throw new Error("Unable to find ERC20 contract for " + token1);
-  }
-
-  const token1Reserves = toDecimal(pair.getReserves().value1, token1Contract.decimals());
-  const token1Rate = getUSDRateUniswapV2(token1, pairAddress, blockNumber);
-  const token1Value = token1Reserves.times(token1Rate);
-  log.debug("token1: reserves = {}, rate = {}, value: {}", [
-    token1Reserves.toString(),
-    token1Rate.toString(),
-    token1Value.toString(),
-  ]);
-
-  const totalValue = token0Value.plus(token1Value);
-  log.info("Total value of pair {} is {}", [pairAddress, totalValue.toString()]);
-  return totalValue;
-}
-
-/**
- * Determines the value of the given balance
- * of a liquidity pool.
- *
- * @param lpBalance
- * @param pairAddress
- * @param blockNumber
- * @returns
- */
-export function getUniswapV2PairValue(
-  lpBalance: BigInt,
-  pairAddress: string,
-  blockNumber: BigInt,
-): BigDecimal {
-  const pair = getUniswapV2Pair(pairAddress, blockNumber);
-  if (!pair) {
-    throw new Error(
-      "Cannot determine discounted value as the contract " + pairAddress + " does not exist yet.",
-    );
-  }
-
-  const lpValue = getUniswapV2PairTotalValue(pairAddress, blockNumber);
-  const poolTotalSupply = toDecimal(pair.totalSupply(), 18);
-  const poolPercentageOwned = toDecimal(lpBalance, 18).div(poolTotalSupply);
-  const balanceValue = poolPercentageOwned.times(lpValue);
-  log.info("Value for pair {} and balance {} is {}", [
-    pairAddress,
-    lpBalance.toString(),
-    balanceValue.toString(),
-  ]);
-  return balanceValue;
-}
-
-/**
- * Determines the value of the given balance
- * of a liquidity pool between OHM and a USD
- * stablecoin.
- *
- * @param lpBalance
- * @param pairAddress
- * @param blockNumber
- * @returns
- */
-export function getOhmUSDPairValue(
-  lpBalance: BigInt,
-  pairAddress: string,
-  blockNumber: BigInt,
-): BigDecimal {
-  const pair = getUniswapV2Pair(pairAddress, blockNumber);
-  if (!pair) {
-    throw new Error(
-      "Cannot determine discounted value as the contract " + pairAddress + " does not exist yet.",
-    );
-  }
-
-  const ohmReserves = pair.getReserves().value0;
-  const secondTokenReserves = pair.getReserves().value1;
-  const poolTotalSupply = toDecimal(pair.totalSupply(), 18);
-  const poolPercentageOwned = toDecimal(lpBalance, 18).div(poolTotalSupply);
-
-  const ohmValue = toDecimal(ohmReserves, 9).times(getBaseOhmUsdRate(blockNumber));
-
-  // Total value in USD is ohmValue + balance of USD stablecoin
-  // TODO support for price lookup
-  const lpValue = ohmValue.plus(toDecimal(secondTokenReserves, 18));
-
-  return poolPercentageOwned.times(lpValue);
 }
